+++ patientMatcher/match/phenotype_matcher.py
@@ -10,7 +10,7 @@ from patientMatcher.parse.patient import disorders_to_omim, features_to_hpo
 from patientMatcher.resources import path_to_hpo_terms, path_to_phenotype_annotations
 from patientMatcher.server.extensions import diseases as diseases_extension
 from patientMatcher.server.extensions import hpo as hpo_extension
-from patientMatcher.utils.hpo import HPOIC
+from patientMatcher.server.extensions import hpoic
 
 LOG = logging.getLogger(__name__)
 
@@ -33,14 +33,10 @@ def match(database, max_score, features, disorders):
     omim_terms = []
     query_fields = []
 
-    hpoic = None
-    hpo = None
-
     if features:  # at least one HPO term is specified
         hpo_terms = features_to_hpo(features)
         # compare against all cases which also have features (HPO terms)
         query_fields.append({"features": {"$exists": True, "$ne": []}})
-        hpoic = HPOIC(hpo_extension, diseases_extension)
 
     if disorders:  # at least one OMIM term was provided
         omim_terms = disorders_to_omim(disorders)